fix: atulização de emprestimo

+++ src/main/java/com/senai/gerenciamento_epis/Repo/EmprestimoRepository.java
@@ -3,5 +3,5 @@ package com.senai.gerenciamento_epis.Repo;
 import com.senai.gerenciamento_epis.Entity.EmprestimoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmprestimoRepository extends JpaRepository<EmprestimoEntity, Integer> {
+public interface EmprestimoRepository extends JpaRepository<EmprestimoEntity, Long> {
 }
